renamed editable field to inline cell in ui folder
