+++ front/src/modules/activities/editable-fields/components/ActivityAssigneeEditableFieldEditMode.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
 import { ActivityAssigneePicker } from '@/activities/components/ActivityAssigneePicker';
-import { useInlineCell } from '@/ui/editable-field/hooks/useInlineCell';
+import { useInlineCell } from '@/ui/inline-cell/hooks/useInlineCell';
 import { Activity, User } from '~/generated/graphql';
 
 const StyledContainer = styled.div`
