--- front/src/modules/ui/editable-field/contexts/EditableFieldMutationContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const EditableFieldMutationContext = createContext<any>(undefined);
