--- front/src/modules/ui/editable-field/types/EditableFieldHotkeyScope.ts
@@ -1,3 +1,0 @@
-export enum EditableFieldHotkeyScope {
-  EditableField = 'editable-field',
-}
