+++ front/src/modules/ui/inline-cell/components/InlineCellContainer.tsx
