+++ front/src/modules/ui/inline-cell/contexts/InlineCellMutationContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const InlineCellMutationContext = createContext<any>(undefined);
