+++ front/src/modules/ui/inline-cell/property-box/components/PropertyBox.tsx
