+++ front/src/modules/ui/inline-cell/states/customEditHotkeyScopeForFieldScopedState.ts
