+++ front/src/modules/ui/inline-cell/states/isInlineCellInEditModeScopedState.ts
