+++ front/src/modules/ui/inline-cell/types/InlineCellHotkeyScope.ts
@@ -1,0 +1,3 @@
+export enum InlineCellHotkeyScope {
+  InlineCell = 'inline-cell',
+}
